Slice 1c.1: PolicyKind trait + policy_canonical module extraction (#302)
Pure refactor — no behavioral change. Extracts the egress canonical
parser + cache from policy_fetcher.rs into a new policy_canonical/
module behind a PolicyKind trait. The trait makes the cosign/OCI
verification pipeline generic over per-kind canonical-form rules,
unblocking Slices 1c.2-1c.5 (tools, inference, memory, mcp-tools).

Byte-identical egress behavior preserved — fetch_and_verify keeps
the same signature/return type via a thin wrapper around
fetch_and_verify_generic::<EgressKind>. validate_ref_shape and
verify_via_sigstore are now generic over K: PolicyKind, and
pull_artifact_bytes takes the per-kind media_type as a parameter.

VerifiedAllowlist, CanonicalEndpoint, and EGRESS_ALLOWLIST_V1_MEDIA_TYPE
moved to controller::policy_canonical::egress; policy_fetcher
re-exports them so existing call sites in this crate compile
unchanged. The per-kind cache lives in egress.rs via a static
OnceLock<Mutex<HashMap<…>>> helper, isolating cache slots per kind.

Tests: 597 controller tests pass (+ 17 new egress.rs unit tests);
existing test module refactored to call EgressKind::cache_clear
and crate::policy_canonical::egress::parse directly. clippy
-D warnings clean, fmt clean.
